Downloader branch
Created Downloader class for downloading datasets. It is consisted of BaseDownloader class that all downloaders should extend. In the future there is a proposition to add a factory method to base class. 
Another class is HttpDownloader that all downloaders that use HTTP protocol should inherit and there is also a simple implementation that uses HTTP protocol without authentication or autorization. 
Download is managed in chunks so that there shouldn't be problems with RAM. 

Added Utility class in takepod.storage directory. It contains function for copying data from one file like object to another. Implementation is based on shutils.copyfileobj function but adds progress bar feature. 

Added functionalities are tested in test_downloader.py module.