Added OoV flagging to lemmatizer
The lemmatizer now includes an optional boolean flag "none_if_oov" which will,
when True, make the method return None if the word is out-of-vocabulary (OoV).

While modifying this, it was neccessary to change the way the wrapper
"capitalize_target_like_source" forwards keyword arguments. This was done in a
backwards-compatible way, so it should not be a problem.

Finally, a test was added for the OoV flag.

Added kwargs to stem_word().

The method stem_word() is now compliant with the changes in utils.py - it can
handle kwargs parameters added in ebaff4da48adc3c6c93ced9808a88307cc1d2b9f.

Minor doc and lint updates.

Updated the multi-line comment in util.py to reflect the new support for kwargs
Deleted blank line in the lemmatizer test to make flake8 happy.

Renamed 'asdfasdf' to 'outofvocabularyword'